% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWIOTv2.R
\name{getWIOTv2}
\alias{getWIOTv2}
\title{Download a WIOT from the second version of the WIOD project}
\usage{
getWIOTv2(period = 2010, format = "wide", version = "October16",
  as.DT = TRUE)
}
\arguments{
\item{period}{the year of the WIOT}

\item{format}{a character string to specify the needed format,
either "wide" (default), "long" or "list". See Details for more
information}

\item{version}{for (possible further) revisions. Currently not
used.}

\item{as.DT}{a logical value: Should the "wide" or "long" formats
be returned as data.tables? Default is TRUE.}
}
\value{
the function returns either a data.table (or data.frame
    when as.DT = FALSE) for the chosen formats "wide" or "long", or
    a list for format "list"
}
\description{
This function allows you to download a World Input-Output Table
(WIOT) from the second generation of the WIOD project
(www.wiod.org) for a given year. Additionally you can specify among
three possible formats (wide, long, list).
}
\details{
If format = "wide" is chosen, the data is returned in a
    wide data.table. This format is the format you know from the
    excel-sheets (available from www.wiod.org).

    The format option "long" returns a reshaped/melted
    data.table.

    Finally, format = "list" returns a list of four
    matrices/vectors: 1) matrix Z, the square (2464 x 2464) matrix
    of intermediates, 2) matrix F, the final demand block (2464 x
    220), 3) vector x, gross output per country sector, 4) vector
    v, value added per country sector, 5) matrix va.block, the
    valuation block (i.e. taxes, adjustments for exports,
    international trade margins, etc.) and 6) 1-element vector
    year, specifying the year of the WIOT.
}
\examples{
wiot.2005 <- getWIOTv2(period = 2005, format = "wide")

## get all WIOTs at once
## wiod.list <- list()
## for(this.year in 2000:2014) {
##   wiod.list[[as.character(this.year)]] <- getWIOTv2(period = this.year,
##                                                     format = "list")
##}

## decompose the 2005 WIOT using the WWZ decomposition
## library(decompr)
## wiot.2005 <- getWIOTv2(period = 2005, format = "list")
## industries <- 1:56
## countries <- unique(substring(names(wiot.2005$x), 1, 3))
## decomp <- decomp(x = wiot.2005$Z,
##                  y = wiot.2005$F,
##                  k = countries,
##                  i = industries,
##                  o = wiot.2005$x,
##                  V = wiot.2005$v,
##                  method = "wwz",
##                  verbose = TRUE)
}
\author{
Oliver Reiter
}
\references{
Please cite Timmer, M. P., Dietzenbacher, E., Los, B.,
    Stehrer, R. and de Vries, G. J. (2015),
    "An Illustrated User Guide to the World Input–Output Database: the Case of Global Automotive Production". Available
    at \url{http://dx.doi.org/10.1111/roie.12178}
}
\seealso{
\url{www.wiod.org}
}
\keyword{data,}
\keyword{input-output,}
\keyword{wiod}

